requirements.txt
requests>=2.31.0
geopy>=2.4.0
plyer>=2.1.0
Pillow>=10.0.0
pydub>=0.25.1
pystray>=0.19.0
beautifulsoup4>=4.12.0
psutil>=5.9.0
